Added add_comm_info routine but use of comms module is a bad solution
In the present communications model, which uses point-to-point communicator objects set up separately by manager and worker, the communicators are not created until after manager and workers have forked, and so global comms, while possible, would require sync between manager and workers, which seems to defeat the point of separation. In the alternative global comms object approach, this comms object would be set up before forking manager/worker and such comms could be done at that point. In the present case, which perhaps has the benefit of being possible to implement in any point-to-point communications medium) this will have to be done using provided communicator information in libE before comms module is created (and hence before the manager/worker fork).